Add Multer for file upload handling
Integrated Multer to handle file uploads in the `upload-file` endpoint. Modified relevant functions to support folder path manipulation and updated the main function to use a POST request for file uploading. Removed extraneous commented-out code for better clarity.

+++ functions/components/deleteFile.js
@@ -1,14 +1,16 @@
 const {shareClient} = require("../AzureCredentials");
 
 async function deleteFile(folderPath, fileName) {
-    let directoryPath = "CLIENTS/Other Clients/MCO";
+    if (folderPath !== undefined && fileName !== undefined) {
+        let directoryPath = "CLIENTS/Other Clients/MCO";
 
-    let fullPath =  directoryPath + "/" + folderPath;
-    const directoryClient = shareClient.getDirectoryClient(fullPath);
-    const fileClient = directoryClient.getFileClient(fileName);
+        let fullPath = directoryPath + "/" + folderPath;
+        const directoryClient = shareClient.getDirectoryClient(fullPath);
+        const fileClient = directoryClient.getFileClient(fileName);
 
-    await fileClient.delete();
-    console.log("Deleted file");
+        await fileClient.delete();
+        console.log("Deleted file");
+    }
 }
 
 deleteFile().catch(e => {
