+++ functions/components/deleteFolder.js
@@ -4,11 +4,13 @@ async function deleteFolder(folderPath) {
     let directoryPath = "CLIENTS/Other Clients/MCO";
 
     let fullPath =  directoryPath + "/" + folderPath;
-    console.log(fullPath)
-    const directoryClient = shareClient.getDirectoryClient(fullPath);
 
-    await directoryClient.delete();
-    console.log("Deleted directory");
+    if (folderPath !== undefined) {
+        const directoryClient = shareClient.getDirectoryClient(fullPath);
+
+        await directoryClient.delete();
+        console.log("Deleted directory");
+    }
 }
 
 deleteFolder().catch(e => {
